+++ src/Commands/ControllerMakeCommand.php
@@ -2,6 +2,7 @@
 
 namespace Nwidart\Modules\Commands;
 
+use Illuminate\Support\Str;
 use Nwidart\Modules\Support\Config\GenerateConfigReader;
 use Nwidart\Modules\Support\Stub;
 use Nwidart\Modules\Traits\ModuleCommandTrait;
@@ -88,6 +89,7 @@ class ControllerMakeCommand extends GeneratorCommand
     {
         return [
             ['plain', 'p', InputOption::VALUE_NONE, 'Generate a plain controller', null],
+            ['api', null, InputOption::VALUE_NONE, 'Exclude the create and edit methods from the controller.'],
         ];
     }
 
@@ -96,9 +98,9 @@ class ControllerMakeCommand extends GeneratorCommand
      */
     protected function getControllerName()
     {
-        $controller = studly_case($this->argument('controller'));
+        $controller = Str::studly($this->argument('controller'));
 
-        if (str_contains(strtolower($controller), 'controller') === false) {
+        if (Str::contains(strtolower($controller), 'controller') === false) {
             $controller .= 'Controller';
         }
 
@@ -119,15 +121,19 @@ class ControllerMakeCommand extends GeneratorCommand
     }
 
     /**
-     * Get the stub file name based on the plain option
+     * Get the stub file name based on the options
      * @return string
      */
     private function getStubName()
     {
         if ($this->option('plain') === true) {
-            return '/controller-plain.stub';
+            $stub = '/controller-plain.stub';
+        } elseif ($this->option('api') === true) {
+            $stub = '/controller-api.stub';
+        } else {
+            $stub = '/controller.stub';
         }
 
-        return '/controller.stub';
+        return $stub;
     }
 }