+++ tests/LumenModuleTest.php
@@ -3,7 +3,6 @@
 namespace Nwidart\Modules\Tests;
 
 use Nwidart\Modules\Json;
-use Nwidart\Modules\Module;
 
 class LumenModuleTest extends BaseTestCase
 {
@@ -12,7 +11,7 @@ class LumenModuleTest extends BaseTestCase
      */
     private $module;
 
-    public function setUp()
+    public function setUp(): void
     {
         parent::setUp();
         $this->module = new LumenTestingModule($this->app, 'Recipe Name', __DIR__ . '/stubs/valid/Recipe');